# Conflicts:
#	Source/AGXUnreal/Private/AGX_Simulation.cpp
#	Source/AGXUnreal/Private/Shapes/AGX_ShapeComponent.cpp
#	Source/AGXUnreal/Public/AGX_Simulation.h
#	Source/AGXUnrealBarrier/Private/AGXBarrierFactories.h
#	Source/AGXUnrealBarrier/Private/SimulationBarrier.cpp
#	Source/AGXUnrealBarrier/Public/RigidBodyBarrier.h
#	Source/AGXUnrealBarrier/Public/SimulationBarrier.h
#	Source/AGXUnrealEditor/Private/AGXUnrealEditor.cpp

+++ Source/AGXUnrealBarrier/Private/Wire/WireBarrier.cpp
@@ -1,16 +1,16 @@
-// Copyright 2024, Algoryx Simulation AB.
+// Copyright 2025, Algoryx Simulation AB.
 
 #include "Wire/WireBarrier.h"
 
 // AGX Unreal includes.
 #include "AGXBarrierFactories.h"
 #include "BarrierOnly/AGXRefs.h"
+#include "BarrierOnly/AGXTypeConversions.h"
 #include "BarrierOnly/Wire/WireNodeRef.h"
 #include "BarrierOnly/Wire/WireParameterControllerPtr.h"
 #include "BarrierOnly/Wire/WireRef.h"
 #include "BarrierOnly/Wire/WireWinchRef.h"
 #include "Materials/ShapeMaterialBarrier.h"
-#include "TypeConversions.h"
 #include "Wire/WireNodeBarrier.h"
 #include "Wire/WireWinchBarrier.h"
 #include "Wire/WireParameterControllerBarrier.h"
@@ -42,18 +42,18 @@ FWireBarrier::~FWireBarrier()
 	// forward declaration, of FWireRef.
 }
 
-void FWireBarrier::SetRadius(float Radius)
+void FWireBarrier::SetRadius(double Radius)
 {
 	check(HasNative());
 	const agx::Real RadiusAGX = ConvertDistanceToAGX(Radius);
 	NativeRef->Native->setRadius(RadiusAGX);
 }
 
-float FWireBarrier::GetRadius() const
+double FWireBarrier::GetRadius() const
 {
 	check(HasNative());
 	const agx::Real RadiusAGX = NativeRef->Native->getRadius();
-	const float Radius = ConvertDistanceToUnreal<float>(RadiusAGX);
+	const double Radius = ConvertDistanceToUnreal<double>(RadiusAGX);
 	return Radius;
 }
